feat(user): change user ID type from int to long across endpoints, DTOs, and services

+++ Backend/app/Core/Interfaces/IUserRepository.cs
@@ -12,13 +12,13 @@ public interface IUserRepository
     // CRUD stands for Create, Read, Update, and Delete – basic operations used to add, fetch, modify, and remove data.
 
     Task<IEnumerable<User>> GetAllUsersAsync(); //
-    Task<User?> GetUserByIdAsync(int id); //
+    Task<User?> GetUserByIdAsync(long id); //
     Task<User?> GetUserByEmailAsync(string email); //
     Task<User?> GetUserByRoleAsync(UserRole role); //
     Task<User?> GetUserByDisplayNameAsync(string displayName); //
     Task<bool> CreateUserAsync(User user); //
-    Task<bool> UpdateUserAsync(int id, User user); //
-    Task<bool> DeleteUserAsync(int id); //
+    Task<bool> UpdateUserAsync(long id, User user); //
+    Task<bool> DeleteUserAsync(long id); //
 
     // New query methods
     Task<IEnumerable<User>> GetUsersByClassAsync(string classname); //
